requirements.txt
# Flask==2.2.5
# redis==4.5.0
# flask-socketio==5.3.6
# flask_jwt_extended==4.5.3
# flask-cors==4.0.0
# pymongo[srv]==4.5.0
# certifi==2025.1.31
# gunicorn==20.1.0
fastapi 
uvicorn 
pymongo 
redis 
python-dotenv 
fastapi-jwt-auth<0.6
pydantic<2

